requirements.txt
python-chess==0.27.3
